pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cen_detect_hor"
dynamic = ["version"]
description = 'Library to characterise tandem repeats in genomic sequences in terms of evolutionary relationships, looking for so called higher-order-repeats. It has been pecifically designed to analyse non-coding centromeric high-repetition regions of chromosomes.'
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
keywords = []
authors = [
  { name = "Miguel Ceriani", email = "miguel.ceriani@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12"
]
dependencies = [
  "biopython>=1.84",
  "editdistance>=0.8.1",
  "matplotlib>=3.9.2",
  "numpy>=2.0.2",
  "scikit-learn>=1.5.2",
  "typing_extensions>=4.12.2"
]

[project.urls]
Documentation = "https://github.com/CENdetectHOR/CENdetectHOR_lib#readme"
Issues = "https://github.com/CENdetectHOR/CENdetectHOR_lib/issues"
Source = "https://github.com/CENdetectHOR/CENdetectHOR_lib"

[tool.hatch.version]
path = "src/cen_detect_hor/__about__.py"
